requirements.txt
google-cloud-aiplatform>=1.38.0
google-cloud-discoveryengine>=0.11.0
google-cloud-storage>=2.10.0
google-adk>=0.1.0
beautifulsoup4>=4.12.0
requests>=2.31.0
langchain>=0.1.0
langchain-google-vertexai>=1.0.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
lxml>=4.9.0
html2text>=2020.1.16
flask>=3.0.0
flask-cors>=4.0.0

